Add files via upload
toggle a switch, the action happens pretty fast on the device, but the loading bar remains for a pretty longer time, tune this delay time
-- done
on loading the switches page of a device, first the old values are loaded, either keep the page empty at the start, or load the page only when you have correct data to populte..If possible, load with last known names, when the fresh data comes, refresh the UI with new data. --done


on toggling the bottom tabs, no need to reload the enter data from API call, the data should be populated from local cache, then we can initiate a bg call, and if there is any new data available, we can populate that. --done


the initial load time is too much, we need to figure out what are the calls happening, and if there are some optional calls, we can as well plan to remove them --done


while setting up a new device, on the wifi populate list, populate the list in the decreasing order of wifi strength. --done

Also, if possible, test the wifi ssid string, and if u feel this is our device WIFI, put this on top (when looking for a device wifi list) or at the bottom when choosing the home wifi network --done


On configuring new device, configuring device loader seems stuck for a very long time..tune this resp time, decreasing the status polling rate --done